+++ src/inspect_ai/util/__init__.py
@@ -2,7 +2,9 @@ from ._concurrency import concurrency
 from ._console import input_screen
 from ._resource import resource
 from ._sandbox import (
+    OutputLimitExceededError,
     SandboxEnvironment,
+    SandboxEnvironmentLimits,
     SandboxEnvironments,
     SandboxEnvironmentSpec,
     SandboxEnvironmentType,
@@ -22,9 +24,11 @@ __all__ = [
     "ExecResult",
     "concurrency",
     "input_screen",
+    "OutputLimitExceededError",
     "resource",
     "subprocess",
     "SandboxEnvironment",
+    "SandboxEnvironmentLimits",
     "SandboxEnvironments",
     "SandboxEnvironmentSpec",
     "SandboxEnvironmentType",