+++ src/inspect_ai/util/_sandbox/environment.py
@@ -110,6 +110,10 @@ class SandboxEnvironment(abc.ABC):
         The current working directory for execution will be the per-sample
         filesystem context.
 
+        Each output stream (stdout and stderr) is limited to 1 MiB. If exceeded, an
+        `OutputLimitExceededError` will be raised. This will not cause the command to be
+        terminated.
+
         Args:
           cmd (str | list[str]): Command or command and arguments to execute.
           input (str | bytes | None): Standard input (optional).
@@ -127,6 +131,8 @@ class SandboxEnvironment(abc.ABC):
             decoding the command output.
           PermissionError: If the user does not have
             permission to execute the command.
+          OutputLimitExceededError: If an output stream
+            exceeds the 1 MiB limit.
         """
         ...
 
@@ -160,6 +166,8 @@ class SandboxEnvironment(abc.ABC):
     async def read_file(self, file: str, text: bool = True) -> Union[str | bytes]:
         """Read a file from the sandbox environment.
 
+        File size is limited to 100 MiB.
+
         Args:
           file (str): Path to file (relative file paths will resolve to the
             per-sample working directory).
@@ -176,6 +184,8 @@ class SandboxEnvironment(abc.ABC):
           PermissionError: If the user does not have
             permission to read from the specified path.
           IsADirectoryError: If the file is a directory.
+          OutputLimitExceededError: If the file size
+            exceeds the 100 MiB limit.
         """
         ...