+++ src/inspect_ai/util/_sandbox/self_check.py
@@ -1,6 +1,13 @@
 from typing import Any, Callable, Coroutine, Generic, Optional, Type, TypeVar
+from unittest import mock
 
-from inspect_ai.util import SandboxEnvironment
+import pytest
+
+from inspect_ai.util import (
+    OutputLimitExceededError,
+    SandboxEnvironment,
+    SandboxEnvironmentLimits,
+)
 
 
 async def check_test_fn(
@@ -32,6 +39,7 @@ async def self_check(sandbox_env: SandboxEnvironment) -> dict[str, bool | str]:
         test_read_file_not_allowed,
         test_read_file_is_directory,
         test_read_file_nonsense_name,
+        test_read_file_limit,
         test_write_file_zero_length,
         test_write_file_space,
         test_write_file_is_directory,
@@ -45,6 +53,9 @@ async def self_check(sandbox_env: SandboxEnvironment) -> dict[str, bool | str]:
         test_cwd_custom,
         test_cwd_relative,
         test_cwd_absolute,
+        test_exec_stdout_is_limited,
+        test_exec_stderr_is_limited,
+        test_exec_limit_exceeded_does_not_terminate_command,
     ]:
         results[fn.__name__] = await check_test_fn(fn, sandbox_env)
 
@@ -139,6 +150,17 @@ async def test_read_file_nonsense_name(
     with Raises(FileNotFoundError) as e_info:
         await sandbox_env.read_file(file, text=True)
     assert "wikipedia" in str(e_info.value)
+
+
+async def test_read_file_limit(sandbox_env: SandboxEnvironment) -> None:
+    file_name = "large.file"
+    await sandbox_env.write_file(file_name, "a" * 2048)  # 2 KiB
+    # Patch limit down to 1KiB for the test to save us from writing a 100 MiB file.
+    with mock.patch.object(SandboxEnvironmentLimits, "MAX_READ_FILE_SIZE", 1024):
+        with Raises(OutputLimitExceededError) as e_info:
+            await sandbox_env.read_file("large.file", text=True)
+        assert "limit of 100 MiB was exceeded" in str(e_info.value)
+    await _cleanup_file(sandbox_env, file_name)
 
 
 async def test_write_file_zero_length(sandbox_env: SandboxEnvironment) -> None:
@@ -270,6 +292,42 @@ async def test_cwd_absolute(sandbox_env: SandboxEnvironment) -> None:
     current_dir_contents = (await sandbox_env.exec(["ls"], cwd=cwd_directory)).stdout
     assert "test_cwd_absolute.file" in current_dir_contents
     await _cleanup_file(sandbox_env, file_name)
+
+
+async def test_exec_stdout_is_limited(sandbox_env: SandboxEnvironment) -> None:
+    output_size = 1024**2 + 1024  # 1 MiB + 1 KiB
+    with pytest.raises(OutputLimitExceededError) as e_info:
+        await sandbox_env.exec(["sh", "-c", f"yes | head -c {output_size}"])
+    assert "limit of 1 MiB was exceeded" in str(e_info.value)
+    truncated_output = e_info.value.truncated_output
+    # `yes` outputs 'y\n' (ASCII) so the size equals the string length.
+    assert truncated_output and len(truncated_output) == 1024**2
+
+
+async def test_exec_stderr_is_limited(sandbox_env: SandboxEnvironment) -> None:
+    output_size = 1024**2 + 1024  # 1 MiB + 1 KiB
+    with pytest.raises(OutputLimitExceededError) as e_info:
+        await sandbox_env.exec(["sh", "-c", f"yes | head -c {output_size} 1>&2"])
+    assert "limit of 1 MiB was exceeded" in str(e_info.value)
+    truncated_output = e_info.value.truncated_output
+    assert truncated_output and len(truncated_output) == 1024**2
+
+
+async def test_exec_limit_exceeded_does_not_terminate_command(
+    sandbox_env: SandboxEnvironment,
+) -> None:
+    file = "limit-exceeded-but-still-running.file"
+    # Write 2 MiB to stdout, then create a file.
+    script = f"""\
+yes | head -c {2 * 1024**2}
+touch {file}
+"""
+    with pytest.raises(OutputLimitExceededError):
+        await sandbox_env.exec(["sh", "-c", script])
+    # Verify that exceeding the output limit did not terminate the command.
+    test_result = await sandbox_env.exec(["test", "-f", file])
+    assert test_result.returncode == 0
+    await _cleanup_file(sandbox_env, file)
 
 
 # TODO: write a test for when cwd doesn't exist